add some tests
