--- otail-web/tests/e2e/example.spec.ts
@@ -1,6 +1,0 @@
-import { test, expect } from '@playwright/test';
-
-test('basic test', async ({ page }) => {
-  await page.goto('/');
-  await expect(page).toHaveTitle(/OTail/);
-});
